Support for code bindings on profiles
Initial support for applying code constraints (as bindings) on profiles.  This includes work around constraints in general to ensure all constraints in the hierarchy are picked up.   In the future, it might be nice to beef up the model to make this easier.

No constraints being applied to extensions yet.