Support for 'Profile Indicators'
There was a need for a file to show profiles and the fixed values within them -- in order to take a resource and "guess" what profile it should adhere to (part of the FHIR Scorecard effort).  This adds support for generating such a file -- although output is turned off by default.