Fix bugs in processing field to field cardinality
Previously, the parent aggregate cardinality was determined by looking up the parent by path.  This did not work correctly with sliced elements (since their are multiple instances of the same path).  Now find parents by ID.

Also fix the code that tries to determine if a give path has a mapping rule.  Previously it tried to match on the whole target, but when the target had commands (e.g., slicing, etc), it didn't work correctly.  Now match on the target, ignoring commands.