Warn when overriding extensible bindings
FHIR allows extensible bindings to be overridden, but only when the new codes do NOT overlap the meaning of any of the old codes.  We can't ensure this programatically, so the best thing we can do is print a warning to make sure the author is aware and it is intentional.

See standardhealth/shr_spec/issues/37.