Log errors/warnings when equal choice option paths
In some cases, a choice may have options of the same type (e.g., CodeableConcept) that only differ in constraints (e.g., different value sets).  In these cases, they'll map to the same element.  Previously we just mapped the first one and skipped the others.  Now we purposefully don't skip the others and log a warning when we detect the situation.  When this results in trying to apply multiple bindings, it will also generate an error (since a required binding can't be overridden).

See standardhealth/shr_spec/issues/36.