More agressive choice mapping
When mapping against choices, try to map to as many as possible by attempting to map the source's value as well.  For example, in MedicationPrescription, mapping to medication[x] should map the Medication element *and* the CodeableConcept (which is the Medication's value).  Prior to this it only mapped the Medication and then stopped.