Support Expanded Choice Value Mappings
When a mapping uses the special "Value" word in its path, and that value is actually a choice, the expander will now expand that into a new mapping rule per choice option.

One result is that we now have to handle the case that if a subclass constrains out some choices, we need to ignore any inherited mappings that might still refer to those choices.