Update to FHIR 3.0.1
Update from FHIR 3.0.0 to FHIR 3.0.1 definitions and ig publisher jar.  This fixes some issues in publishing regarding the resolution of value sets (and likely other issues as well).

This also updates the text of the index page to be more accurate and appropriate (representing FHIR 3.0.1 instead of the previous FHIR 1.8).