Support for slicing w/ IncludesTypeConstraint
Now you can slice on a field that has IncludesTypeConstraints and each includes type will be a slice.

I don't love this code.  It is not perfect.  But... the current codebase doesn't lend itself well to this task.  Now that we know the requirements better, we should consider the value of refactoring the shr-fhir-export (and shr-models) to better fit our needs.

But... for now... this works.