Tentative fix for issue shr-fhir-export#41: String value constraint on Observation generating spurious Warning
Previously a warning was raised when the value of an object based on Observation was a primitive.

There previously was no check for a value of a primitive, so I added one. However, I'm not too familiar with what's supposed to be happening in this space, so I'm marking this as a tentative solution that needs review.

Solution to issue:
https://github.com/standardhealth/shr-fhir-export/issues/41