Simpler approach to supporting recursive definitions in extensions
While the previous approach appeared to work, the fact that createExtension sometimes returned an extension and sometimes returned null was a little confusing and partly concerning -- especially since null didn't necessarily mean the extension was invalid (it was just currently in process).

This new approach is simpler and avoids the issue noted above.  Note that it no longer logs an error if a recursive definition is found, because that's not necessarily an error condition anyway.