Load US Core IG
This commit adds support for loading the US Core IG, but does not yet support USE of the IG definitions.  In addition, it adds logic to allow any definitions (base FHIR or IG) to be found by ID or URL.  Lastly, old no-op test replaced with tests for the definitions loader.