Consider choices when getting FHIR cardinalities
Previously, if a FHIR element had cardinality 1..1, but multiple choices, it would still report its cardinality as 1..1.  This caused problems in mapping, because really, any of the choices might *not* be chosen, so technically, they are each 0..1.

Now, if we detect a choice w/ more than one option, we treat the min as 0 since any one of those options might *not* be chosen.