Fix FindValueBy... and detection of compatible profile types
There was a problem with FindValueByIdentifier and FindValueByPath that was not properly merging down constraints and was not recognizing when a choice was constrained to a single type via a constraint.  This has been fixed.

In addition, a bug that converted refs to Identifiers has also been fixed.

Last, fixed an issue where mapping a field to a profile reference from a different IG wasn't correctly detecting compatibility.  For example, if a PersonOfRecord field was mapped to a property that was profiled to us-core-patient, it complained even though PersonOfRecord is mapped to us-core-patient.  This is fixed now.