Supress unnecessary extension unrolling
When applying contraints, code that searches for the "constrainable element" was being applied before we even checked if there was a relevant constraint to apply.  This caused extensions to be unnecessarily unrolled.  This is fixed by only searching for the constrainable element if we know there is indeed something to be constrained.