Log warnings for possibly problematic recursive lookups
In some cases, we look up a profile or extension and copy its contents into another profile, or check if it is a no-diff profile.  If the profile/extension we are looking up is currently in the middle of processing (due to a recursive relationship), the results may be unexpected.  While we would ideally protect against this issue entirely, that will require significant work.  Instead, we log a warning in those cases.  Currently, this warning is NOT triggered by current SHR definitions.

Also fixed some error numbers in some log statements.