Make slice IDs local to each structure definition
Instead of having a global slice id counter, make it local to each structure def (e.g., each structuredef restarts at 1).

This is primarily to make it easier to compare two sets of SHR definitions.  Before this change, the addition of one slice in one profile could result in *many* profiles being different because it incremented the global counter.  Now there will only be diffs for structuredefs that changed in a more significant way.