Models: Every element should have a base property
There is a bug in the current publisher that requires every element to define a base.  This is a rule in R4, but for some reason is being enforced on R3 structure definitions.

Since it's not WRONG to have a base, we'll get around this by just doing what the publisher wants, even if it is not a string R3 requirement.

See: https://chat.fhir.org/#narrow/stream/IG.20creation/topic/IGPub.20Error.3A.20Models.20-.20element.2Ebase.20cannot.20appear.2E.2E.2E