No-diff element types shouldn't create differential
If a mapping doesn't cause a difference to an element, then that element
shouldn't go into the differential.  Prior to this commit, that could
happen when the original specified a profile and the thing mapped
turned out to be a no-diff profile.  For example, Range specifies low
and high as SimpleQuantities.  This should make Range a no-diff profile
but the SimpleQuantity profiles were causing a differential to be made
even when it was not nessecary.

See: https://github.com/standardhealth/shr-cli/issues/116