Create Basic Resource Type Code System
Profiles on Basic resources are distinguished via codes.  Previously we would assign a profile a code, but we didn't actually create the code system we associated it to.  Now we create a Basic Resource Type code system that has codes for all of our profiles on Basic.

Fixes hl7/us-breastcancer#6