Support optional "compact" models
If the config.implemenationGuide.compactModels is set to true, then the logical models are compacted.

This means that any elements whose type is an element that is just a single value will be updated to point directly to that value.

For example, if a ClinicalStatus element (field) has type ClinicalStatus-model, and the ClinicalStatus-model model just has a value that is a Coding, then the ClinicalStatus element will be updated to be type Coding.