Fix issue w/ specialized types and explicit choices
Specialized types can fit into other types, but when you make a choice explicit you must pick one of the existing types, even if it is not specialized.

For example, Count specializes Quantity -- but if you map a Count to value[x], you cannot make it valueCount -- it must be valueQuantity.

Fixes HL7/us-breastcancer#11