Fix resolve func in logical models to use cache
The resolve function use to re-export the model every time it was
called, even if the model had been previously exported. This could
cause a bug when compacting models that would overwrite a compacted
model with an un-compacted one (when resolve re-exports it). In the
shimi branch, this was seem with Quantity's 'comparator' field.