Squashed commits from cmoesel to further support toFHIR/fromFHIR:
Avoid duplicate SHR maps from element.mapping

Prior implementation sometimes pushed the same SHR map onto an element multiple times.  This only adds an SHR map if it doesn't already exist.

Remove redundant and partially incorrect mapping insertion

SHR mapping rules are inserted as part of the field-to-field processing.  The SHR mapping insertion happening in `mappingToProfile` is redundant.  Furthermore, it is buggy, in that it inserts only the first part of the path, ignoring nested parts of the path.  This code can be safely removed.