Squashed commits from cmoesel to fix issues w/ toFHIR/fromFHIR:
Add base to extension slices

Extension slices can and should have a base defined (which should always be 0..*).

Record mappings for primitives too

When unrolling elements into a profile, remove SHR mappings

Since the SHR mappings recorded in a profile are always relative to the root of the profile, they don't work when a profile is expanded into a specific path of another profile.  For this reason, remove mappings when this happens.

Set the base on slices

When we create slices, set the base so tooling can easily see the original cardinality.

Add base when setting cardinality on FHIR elements