Fix slicename detection for DSTU2
Since DSTU2 uses the same field for element names and slice names, you need to do some work to figure out if a name is really a *slice* name or not.  Prior to this fix, the algorithm could be wrong if there were slices on the same path in a different sub-tree of the elements.