Fix type selection for slices
When we create slices, we clone the base element that is getting sliced.  If this element already has types that are marked as selected, then the previous logic would never narrow them down further in the slice.

Now we have to notion of a type priority so that when the slice is cloned, the current selections are given lower priority (actually represented by higher numbers) -- so that if a constraint modifies a type in a slice, the modification will override the cloned type selections -- and only it will be preserved.

This also re-uses some cleanup logic in the extension exporter since I noticed that some of our special properties weren't getting cleaned up there.