DSTU2: Instead of removing ids, rewrite them
Removing ids makes the published profiles strange since the publisher copies the `name` to the `id` -- so we end up w/ some elements having ids and others not.

This rewrites the `id`s so they don't have invalid characters.  It still suffers in that the id might be longer than allowed (64 characters).

We also remove any element names that aren't strictly slice names.