Add (hidden) ReleaseHeader to all pages
IG Publisher now wants every page to contain special markup where the "release header" should be inserted.

This only applies to HL7-published IGs, so we implement it to get rid of the errors, but use CSS to hide it for now (since it contains a broken history link).  Eventually we should make it a configuration parameter.