v5.14.0-beta.5 (FHIR R4 Support)
Refactored extension generation

Instead of using an extension template, pull the extension definition from the FHIR definitions and modify it as appropriate to represent an explicit extension.  This aligns more closely w/ how we do profiles -- but also addresses several previously undiscovered bugs related to different FHIR versions.

This work was initially started after a report that an extension for a data element with a value `Money` was generating a complex extension instead of an extension with `valueMoney`.

Update IG Publisher and US Core Definitions