Fix bug that over-agressively compacted snapshot elements
The algorithm to compact "unneeded" snapshot elements did not take into account the use case of a profile based on another profile -- and therefore compacted out some elements that the parent profile introduced.  Now we will not compact any elements that are in the parent profile snapshot; as we assume they must be required.